src/requirements.txt
emoji==0.5.1
numpy==1.16.2
pandas==0.24.1
flask_restplus==0.12.1
Unidecode==1.0.23
matplotlib==3.0.2
seaborn==0.9.0
Flask==1.0.2
langdetect==1.0.7
spacy==2.0.18
cassandra_driver==3.17.1
scikit_learn==0.21.1
